fastapi_ugc/requirements.txt
fastapi==0.96.0
orjson==3.8.5
pydantic==1.9.0
uvicorn==0.12.2
uvloop==0.17.0
gunicorn==20.1.0
python-dotenv==0.21.0
python-jose==3.3.0
aiokafka==0.8.0
requests==2.31.0
pytest==7.3.1
sentry-sdk[fastapi]==1.25.1
python-logstash==0.4.6
motor==3.1.2
